% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simstudy_edd.R
\name{est_edd_all_methods}
\alias{est_edd_all_methods}
\title{Estimates the edd for a given change for all methods.}
\usage{
est_edd_all_methods(train_obj, n_max, n_sim, edd_file, kappa, p, w, p0s,
  r_pca, threshold_settings, mu = NULL, sigma = NULL,
  rho_scale = NULL)
}
\arguments{
\item{train_obj}{A list returned from \code{\link{get_training_sets}}.}

\item{n_max}{The maximum number of monitoring samples.}

\item{n_sim}{The number of run length simulations to estimate the EDD.}

\item{edd_file}{A string specifying the file to print results to.}

\item{kappa}{The change-point.}

\item{p}{A number between 0 and 1. The proportion of affected streams.}

\item{w}{The window length.}

\item{p0s}{A vector of p0-values to be used in the mixture procedure.}

\item{r_pca}{A vector of the numbers of projections to keep in Max and Min PCA.}

\item{threshold_settings}{A list with named elements 'n' and 'alpha', giving
the settings used to control the probability of false alarm.}

\item{mu}{A numeric specifying the change in mean.}

\item{sigma}{A numeric specifying the change in standard deviation.}

\item{rho_scale}{A numeric specifying the change in correlation.}
}
\description{
Description
}
\details{
Simulates \code{n_sim} run lengths and calculates the EDD for all
methods (Mixture, Min PCA, Max PCA and TPCA) together with other summary
statistics of the run length distribution. These values are not returned,
but printed to the input \code{edd_file}. The input \code{kappa}, \code{p},
\code{mu}, \code{sigma} and \code{rho_scale} constitute the change scenario
to estimate the EDD for. \code{p0s} is a vector of parameters for the the
mixture procedure, while \code{r_pca} is a vector of the number of projections
to retain in Max and Min PCA. The chosen projections for TPCA is
found from the file axes-file created during the threshold-finding phase.
}
