% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_threshold.R
\name{est_tpca_threshold}
\alias{est_tpca_threshold}
\title{Estimating the threshold for tpca changepoint detection}
\usage{
est_tpca_threshold(x, axes, arl, rel_tol = c(0.2, 0.1, 0.05, 0.025),
  alpha = 0.05)
}
\arguments{
\item{x}{The d x m training data matrix, where d is the dimension of the data
and m the number of training samples, or a list(d = ., m = .) describing
the dimensions of normal training data to be drawn.}

\item{axes}{Indices of the principal axes to be used in simulations.}

\item{rel_tol}{A vector with the sequence of relative errors allowed at each
step towards convergence in the algorithm. See details.}

\item{alpha}{1 - alpha is the confidence level.}
}
\value{
An estimate of the average run length.
}
\description{
Description
}
\details{
\code{rel_tol} and \code{alpha} governs the number of simulations used in
each step of the algorithm towards a more and more certain estimate.
At each step, the number of simulations is chosen so that
\code{[(1 - rel_tol)arl, (1 + rel_tol)arl]} approximately covers the true
average run length at confidence level \code{alpha}.
For example, when the last \code{rel_tol} is 0.025, it means that the final
estimated threshold corresponds to an average run length of approximately
\code{arl} +- 0.025 * \code{arl} at confidence level \code{alpha}
The algorithm should start with a large relative error tolerance,
and then narrow it down for the quickest convergence.
}
