% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generators.R
\name{gen_train}
\alias{gen_train}
\title{Generate multivariate normal training data.}
\usage{
gen_train(d, m, seed, alphad = 1, return_all = FALSE)
}
\arguments{
\item{d}{The data dimension}

\item{m}{The number of training samples.}

\item{seed}{The seed to be used.}

\item{alphad}{The value of alpha_d in \code{\link{tpca::rcor_mat}}. A high
value means low correlations and vice versa.}

\item{return_all}{A logical value. If FALSE: Only returns the data matrix x.
If TRUE: Returns a list with 'x', 'mu', 'Sigma' and 'nr', where
'mu' is the mean vector used, 'Sigma' is the covariance matrix and 'nr' indicates
which seed that was used.}
}
\value{
See the argument return_all.
}
\description{
Generate multivariate normal training data.
}
