% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generators.R
\name{get_training_sets}
\alias{get_training_sets}
\title{Generate multivariate normal training data.}
\usage{
get_training_sets(n_sets, d, m, return_all = TRUE)
}
\arguments{
\item{n_sets}{The number of training sets to be generated.}

\item{d}{The data dimension.}

\item{m}{The number of training samples.}

\item{return_all}{A logical value. If FALSE: Only returns the data matrix x.
If TRUE: Returns a list with 'x', 'mu', 'Sigma' and 'nr', where
'mu' is the mean vector used, 'Sigma' is the covariance matrix and 'nr' indicates
which seed that was used.}
}
\value{
A list of training sets from \code{\link{gen_train}}.
The argument return_all determines the type of each list element.
}
\description{
Generate multivariate normal training data.
}
