% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_figures.R
\name{ggplot_edd_conf_int}
\alias{ggplot_edd_conf_int}
\title{Plot EDD results with confidence intervals.}
\usage{
ggplot_edd_conf_int(n, alpha, m, d, change_type, change_param,
  ci_alpha = 0.05, cov_mat_nr = NULL, cov_mat_type = NULL,
  summary_func = mean, ylim = c(0, 300), extension = "png")
}
\arguments{
\item{change_type}{Either 'mean', 'sd' or 'cor'}

\item{change_param}{A numeric specifying the change size for the chosen change_type.}
}
\description{
See the documentation for \code{\link{multiplot_edd_summary}} for an explanation
of the arguments. The additional ones are described here.
WARNING: Before running this function
your must either have run \code{run_simstudy} yourself, or have downloaded
the results/ and thresholds/ directories from https://github.com/Tveten/tpcaMonitoring
and put them in your working directory.
}
