% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_figures.R
\name{multiplot_edd_summary}
\alias{multiplot_edd_summary}
\title{Plot EDDS results.}
\usage{
multiplot_edd_summary(n = 100, alpha = 0.01, m = 200, d = 100,
  cov_mat_nr = NULL, cov_mat_type = NULL, change_params = NULL,
  summary_func = mean, ylim = c(0, 300))
}
\arguments{
\item{n}{The monitoring length the probability of false alarm is controlled for.}

\item{alpha}{The probability of false alarms}

\item{m}{The number of training samples.}

\item{d}{The data dimension.}

\item{cov_mat_nr}{The number given the covariance matrix when generated through \code{\link{gen_train}}}

\item{cov_mat_type}{Either 'low' or 'high', to get results only for the low
correlation or high correlation matrices.}

\item{summary_func}{The function to summarize the EDDs over the cov_mat_types.}

\item{ylim}{A vector with lower and upper bounds on the EDD values shown in the plots.}
}
\description{
\code{multiplot_edd_summary} creates a grid of plots with results from
running \code{\link{run_simstudy}}
}
\examples{
# Plot of the results for low correlation matrices:
multiplot_edd_summary(cov_mat_type = 'low')

# Plot of the results for low correlation matrices:
multiplot_edd_summary(cov_mat_type = 'high')

# Plot of the EDD results for a single covariance matrix/training set:
multiplot_edd_summary(cov_mat_nr = 5)

}
