% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simstudy.R
\name{run_simstudy}
\alias{run_simstudy}
\title{Reproduce simulation study with D = 100.}
\usage{
run_simstudy()
}
\description{
This function reproduces the entire simulation study in
"Online Detection of Sparse Changes in High-Dimensional Data Streams Using Tailored Projections"
for D = 100 (Section 4). On an average computer it takes about a week to run on four cores.
Thus, the files with the results of running \code{run_simstudy} used in the paper
are included in the package or found at github.com/Tveten/tpcaMonitoring.
}
\details{
The following directories are created in the current directory:
./thresholds,
./thresholds/axes,
.thresholds/threshold_results,
./results,
./results/figures.
Files created to save information about thresholds are saved in the threshold
directory, while EDD results are stored in "results".
The results we obtained by running \code{run_simstudy} are stored in the
same structure of directories on Github.
}
