% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_threshold.R
\name{tpca_arl}
\alias{tpca_arl}
\title{Estimates average run length (ARL) for monitoring by tpca}
\usage{
tpca_arl(x, threshold, axes, n, n_sim)
}
\arguments{
\item{x}{The d x m training data matrix, where d is the dimension of the data and m the number of training samples, or a list(d = ., m = .) describing the dimensions of normal training data to be drawn.}

\item{threshold}{A numeric specifying the threshold value for when a change
is declared.}

\item{n}{The number of observations to monitor for an estimate of the ARL.
See details.}

\item{n_sim}{The number of simulations to base the estimate on.}

\item{d}{The dimension of the original data stream.}

\item{r}{The dimension of the reduced data stream, i.e.
the number of principal axes to project onto.}

\item{m}{The size of the training set.}

\item{w}{The window size. Number of recent time-points to consider for a
change.}
}
\value{
An estimate of the average run length.
}
\description{
Description
}
\details{
Details
}
